Add files via upload
These are the implementations of the tasks:
Task 1:
Write an application that will calculate and display the value using the formula:
a=4*(b+c-1)/2; b and c are set independently in the code.
Task 2:
The variable n stores a two-digit number. Create a program that calculates and
displaying the sum of digits n. For example: n =26, as a result we should get 8
(2+6)
Task 3:
The variable n stores a three-digit number. Create a program that calculates and
displaying the sum of digits n. For example: n =126, as a result we should get
9 (1+2+6).
Task 4:
The variable n stores a real number with a non-zero fractional part. Create
a program that rounds the number n to the nearest integer and prints the result to
screen.
Task 5:
The variables q and w store two natural numbers. Create a program that prints
display the result of dividing q by w with a remainder. Example program output (for the case
when q stores 21 and w stores 8): 21 / 8 = 2 and 5 remainder.
Task *:
Write a program that will swap the values ​​of integers
variables. Example:
int a = 1; int b = 2;
//code (your decision)
south(a); //will output 2
south(b); //will output 1
Improve the program, use only 2 input variables (a,b)