Add files via upload
This is the implementation of these tasks
Task 1:
Write a program that will take a number from the console as input and output
will display a message whether the number is even or not. To determine the parity of a number
use the operation of obtaining the remainder of division (the operation looks like this: '% 2').
Task 2:
For the entered number t (temperature outside), output: If t>–5, then output “Warm”.
If –5>= t > –20, then output “Normal”. If –20>= t, then output “Cold”.
Task 3:
Write a program that displays the squares of numbers from 10 to 20 inclusive.
Task 4:
It is necessary for the program to display the following sequence on the screen:
7 14 21 28 35 42 49 56 63 70 77 84 91 98. Use a while loop in your solution.
Task *:
Write a program where the user enters any positive integer. A
the program sums all numbers from 1 to the number entered by the user. To enter
numbers, use the Scanner class. Make a check so that the user cannot
enter incorrect data.